+++ test/cmark/parsing.lisp
@@ -40,3 +40,65 @@
   "Parsing an entire document yields a large tree structure"
   (let* ((markdown (load-md-from-file "test/sample.md")))
     (finishes (cmark::parse-document markdown))))
+
+
+;;; ---------------------------------------------------------------------------
+(def-suite cmark/parsing/streaming
+  :description "Testing the streaming parser"
+  :in cmark/parsing)
+(in-suite cmark/parsing/streaming)
+
+(test create-and-close-parser
+  "We can create a new parser and close it"
+  (finishes (cmark::close-streaming-parser (cmark::make-streaming-parser))))
+
+(test with-parser-context
+  "We can call the parser context macro"
+  (finishes
+    (cmark::with-streaming-parser (parser)
+      nil)))
+
+(test parser-context-has-parser-binding
+  "We can call the parser context macro"
+  (cmark::with-streaming-parser (parser)
+    (is (typep parser 'cmark::streaming-parser))))
+
+(test feed-parser
+  "We can feed a string to the streaming parser"
+  (cmark::with-streaming-parser (parser)
+    (finishes (cmark::feed-streaming-parser parser "Hello *world*."))))
+
+(test finish-parser
+  "We can finish the streaming parser"
+  (cmark::with-streaming-parser (parser)
+    (cmark::feed-streaming-parser parser "Hello *world*.")
+    (is (typep (cmark::finish-streaming-parser parser) 'cmark::node))))
+
+(test open-parser-not-exhausted
+  "After closing a parser it is exhausted"
+  (cmark::with-streaming-parser (parser)
+    (is-false (cmark::streaming-parser-exhausted-p parser))))
+
+(test closed-parser-exhausted
+  "After closing a parser it is exhausted"
+  (let ((parser (cmark::make-streaming-parser)))
+    (cmark::close-streaming-parser parser)
+    (is-true (cmark::streaming-parser-exhausted-p parser))))
+
+(test closing-idempotent
+  "Closing a closed parser does not raise any conditions"
+  (let ((parser (cmark::make-streaming-parser)))
+    (finishes (cmark::close-streaming-parser parser))
+    (finishes (cmark::close-streaming-parser parser))))
+
+(test feed-closed
+  "Feeding a closed parser raises a condition"
+  (let ((parser (cmark::make-streaming-parser)))
+    (cmark::close-streaming-parser parser)
+    (signals error (cmark::feed-streaming-parser parser "Hello *world*"))))
+
+(test finish-closed
+  "Finishing a closed parser raises a condition"
+  (let ((parser (cmark::make-streaming-parser)))
+    (cmark::close-streaming-parser parser)
+    (signals error (cmark::finish-streaming-parser parser))))